+++ packages/docs/src/theme/ReactLiveScope/index.js
@@ -3,11 +3,10 @@ import { useStrataSdks } from '@strata-foundation/react';
 import { BN } from "bn.js";
 import { SplTokenBonding } from "@strata-foundation/spl-token-bonding";
 import { SplTokenCollective } from "@strata-foundation/spl-token-collective";
-import { getAssociatedAccountBalance } from "@strata-foundation/spl-utils";
+import { sendMultipleInstructions, getAssociatedAccountBalance, createMetadata, Data  } from "@strata-foundation/spl-utils";
 import { ExponentialCurveConfig, TimeCurveConfig } from "@strata-foundation/spl-token-bonding";
 import { sendAndConfirmRawTransaction, Keypair, PublicKey, Transaction, SystemProgram } from "@solana/web3.js";
-import { TOKEN_PROGRAM_ID, AccountLayout, NATIVE_MINT } from "@solana/spl-token";
-import { Data, sendMultipleInstructions } from "@strata-foundation/spl-utils";
+import { u64, ASSOCIATED_TOKEN_PROGRAM_ID, Token, TOKEN_PROGRAM_ID, AccountLayout, NATIVE_MINT } from "@solana/spl-token";
 import { Numberu32, Numberu64, createInstruction, createNameRegistry, getHashedName, getNameAccountKey, NameRegistryState, NAME_PROGRAM_ID } from "@solana/spl-name-service";
 import BrowserOnly from '@docusaurus/BrowserOnly';
 import { useTokenMetadata, useTokenRef, useTokenBonding, useBondingPricing } from "@strata-foundation/react";
@@ -38,6 +37,12 @@ function Swap(props) {
 
 // Add react-live imports you need here
 const ReactLiveScope = {
+  u64,
+  TOKEN_PROGRAM_ID,
+  ASSOCIATED_TOKEN_PROGRAM_ID,
+  Token,
+  createMetadata,
+  Data,
   NATIVE_MINT,
   getMintInfo,
   getTokenAccount,