+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [Unreleased]
+## [1.3.4](https://github.com/StrataFoundation/strata/compare/v1.3.2...v1.3.4) (2021-12-22)
+
+**Note:** Version bump only for package @strata-foundation/strata
 
 ### Fixed
 
@@ -14,7 +16,13 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 ## Added
 
  * unixTime to pricing so we can price the curve at a given point in time
- 
+
+
+## [1.3.3](https://github.com/StrataFoundation/strata/compare/v1.3.2...v1.3.3) (2021-12-22)
+
+**Note:** Version bump only for package @strata-foundation/strata
+
+* never released
 
 ## [1.3.2](https://github.com/ChewingGlassFund/wumbo-programs/compare/v0.7.0...v1.3.2) (2021-12-15)
 
@@ -117,7 +125,7 @@ Minor version bumping it because the below changes included some interface chang
   - Ability to `updateCollective` to change both the config and authority.
 
 ### Changed
-  - Token collective now lives at TCo1sP6RwuCuyHPHjxgzcrq4dX4BKf9oRQ3aJMcdFry
+  - Token collective now lives at TCo1sfSr2nCudbeJPykbif64rG9K1JNMGzrtzvPmp3y
 
 ### Removed
   - ["token-ref", owner, Pubkey::default] as the primary token ref